Update docker-compose file and disable docker image testing
The docker-compose file has been modified, changing the default argument value syntax. Furthermore, the testing of the Docker image in the GitHub Actions workflow has been disabled. The removed code was in charge of setting up environment variables, setting the image to be tested, starting a Docker compose workflow, and monitoring its operation.